More java5 code and clean up lot of warnings found by Idea
git-svn-id: https://svn.apache.org/repos/asf/continuum/branches/continuum-1.3.x@765340 13f79535-47bb-0310-9956-ffa450edef68

+++ continuum-webapp/src/main/java/org/apache/maven/continuum/web/view/commons/DateCell.java
@@ -51,7 +51,7 @@ public class DateCell
             {
                 Calendar cal = Calendar.getInstance();
 
-                cal.setTimeInMillis( ( (Long) value ).longValue() );
+                cal.setTimeInMillis( (Long) value );
 
                 value = cal.getTime();
             }
